updated to newest gwca version

+++ src/Data/DataSkill.cpp
@@ -8,7 +8,7 @@
 #include "DataSkill.h"
 
 DataSkill::DataSkill(const GW::Constants::SkillID id_, const uint32_t idx_)
-    : id(static_cast<uint32_t>(id_)), idx(idx_), energy_cost(0U), recharge(UINT32_MAX)
+    : id(id_), idx(idx_), energy_cost(0U), recharge(UINT32_MAX)
 {
     const auto* const skill_data = GW::SkillbarMgr::GetSkillConstantData(id);
     if (!skill_data)
